+++ cpg_infra/driver/main_upload_bucket.py
@@ -1,0 +1,12 @@
+# flake8: noqa: ANN401
+"""
+MainUploadBucket named tuple.
+"""
+
+from typing import Any, NamedTuple
+
+
+class MainUploadBucket(NamedTuple):
+    bucket: Any
+    uploaders: list[str]
+    is_dropbox: bool = False
