+++ cpg_infra/driver/sample_metadata_accessor_membership.py
@@ -1,0 +1,12 @@
+# flake8: noqa: ANN401
+"""
+SampleMetadataAccessorMembership named tuple.
+"""
+
+from typing import Any, Iterable, NamedTuple
+
+
+class SampleMetadataAccessorMembership(NamedTuple):
+    name: str
+    member: Any
+    permissions: Iterable[str]
